+++ src/lib.rs
@@ -1,24 +1,30 @@
+mod bookmarks;
 mod error;
 mod parse_search;
 mod parse_tags;
 mod parser;
 mod searcher;
 
-pub use crate::error::XTagError;
-pub use crate::parse_search::compile_search;
-pub use crate::parse_tags::csl_to_map;
-use crate::parser::Rule;
-pub use crate::searcher::Searcher;
-use itertools::Itertools;
-use regex::Regex;
 use std::collections::HashMap;
 use std::path::PathBuf;
 use std::str;
 
+use itertools::Itertools;
+use regex::Regex;
+
+pub use crate::bookmarks::get_bookmark;
+pub use crate::error::{Result, XTagError};
+pub use crate::parse_search::compile_search;
+pub use crate::parse_tags::csl_to_map;
+use crate::parser::Rule;
+pub use crate::searcher::Searcher;
+
+pub type XTags = HashMap<String, Option<String>>;
+
 static XATTR_NAME: &'static str = "user.xtag";
 
 /// Convert map to comma separated list of tag=value pairs
-pub fn map_to_csl(set: &HashMap<String, Option<String>>) -> String {
+pub fn map_to_csl(set: &XTags) -> String {
     set.iter()
         .map(|(tag, value)| match value {
             Some(value) => tag.to_string() + "=" + value,
@@ -28,7 +34,7 @@ pub fn map_to_csl(set: &HashMap<String, Option<String>>) -> String {
 }
 
 /// Get tags for file as map
-pub fn get_tags(path: &PathBuf) -> Result<HashMap<String, Option<String>>, XTagError> {
+pub fn get_tags(path: &PathBuf) -> Result<XTags> {
     let xattrs = xattr::get(path, XATTR_NAME).map_err(|err| XTagError::File(err))?;
     match xattrs {
         Some(value) => {
@@ -42,13 +48,13 @@ pub fn get_tags(path: &PathBuf) -> Result<HashMap<String, Option<String>>, XTagE
 /// Set tags for file from map
 ///
 /// The used utf-8 string format is architecture independent.
-pub fn set_tags(path: &PathBuf, tags: &HashMap<String, Option<String>>) -> Result<(), XTagError> {
+pub fn set_tags(path: &PathBuf, tags: &XTags) -> Result<()> {
     let string = map_to_csl(tags);
     xattr::set(path, XATTR_NAME, &string.as_bytes()).map_err(|err| XTagError::File(err))
 }
 
 /// Delete all tags for file
-pub fn delete_tags(path: &PathBuf) -> Result<(), XTagError> {
+pub fn delete_tags(path: &PathBuf) -> Result<()> {
     match xattr::remove(path, XATTR_NAME) {
         Ok(()) => Ok(()),
         Err(err) if err.to_string().starts_with("No data available") => Ok(()),
@@ -56,12 +62,8 @@ pub fn delete_tags(path: &PathBuf) -> Result<(), XTagError> {
     }
 }
 
-pub fn rename(
-    find: &str,
-    replace: &str,
-    tags: HashMap<String, Option<String>>,
-) -> Result<HashMap<String, Option<String>>, XTagError> {
-    let mut result: HashMap<String, Option<String>> = HashMap::with_capacity(tags.len());
+pub fn rename(find: &str, replace: &str, tags: XTags) -> Result<XTags> {
+    let mut result: XTags = HashMap::with_capacity(tags.len());
     let re = Regex::new(&searcher::expand_regex(find)).map_err(|err| XTagError::Regex(err))?;
     for (key, value) in tags {
         let new_key = re.replace_all(&key, replace).into_owned();
@@ -73,10 +75,11 @@ pub fn rename(
 #[cfg(test)]
 mod tests {
     use super::rename;
+    use super::XTags;
     use std::collections::HashMap;
 
     fn test(key: &str, value: &str, from: &str, to: &str, end_key: &str) {
-        let mut map: HashMap<String, Option<String>> = HashMap::new();
+        let mut map: XTags = HashMap::new();
         map.insert(key.to_string(), Some(value.to_string()));
         let map = rename(from, to, map).unwrap();
         assert_eq!(map.len(), 1);