+++ CHANGELOG.md
@@ -4,15 +4,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## [1.1.0] - 2022-05-27
+### Added
+- Implemented Display for Searcher
+- Implemented bookmark support
+- Defined type aliases Result and XTags
+- Documentation
+
 ## [1.0.0] - 2022-05-16
+### Changed
+- AND now has higher precedence than OR
+- API is not compatible to prior version
+
 ### Added
 - Search can now be precompiled and uses short-circuit evaluation
 - Allows renaming of tags with regex support
 - Supports != operator
-
-### Changed
-- AND now has higher precedence than OR
-- API is not compatible to prior version
 
 ## [0.1.0] - 2022-05-10
 ### Added